+++ CHANGELOG.md
@@ -6,6 +6,10 @@ ByteriX, 2017. All right reserved.
 
 # Versions
 
+## 0.9.21 (24.01.2017)
+##### Bug fixing
+* public protocol BxInputStaticHeight
+
 ## 0.9.20 (23.01.2017)
 ##### Improvements
 * added isEstimatedContent = false for supporting static sizes